requirements.txt
playwright~=1.24.0
PyYAML~=5.4.1
setuptools~=62.3.4